Refactored week 4 challenge with Mike

+++ Stage 1/Week 4/Week4ChallengeBankAccounts/IAnnualAmountEarnedFromInterest.cs
@@ -1,4 +1,4 @@
 interface IAnnualAmountEarnedFromInterest
 {
-    decimal AnnualAmountEarnedFromInterest(); // interface method
+    double AnnualAmountEarnedFromInterest(); // interface method
 }
